Split TypeScript alias import test coverage
